backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
torch>=2.6.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
ultralytics>=8.0.0